resuelve problema en la actualizacion de pagina al usar parametros con use params y carga todas las imagenes locales
