+++ package.json
@@ -4,7 +4,7 @@
   "description": "Academia Geek - Daily Bits",
   "main": "index.js",
   "scripts": {
-    "start": "webpack serve --mode development",
+    "start": "webpack serve --mode development --open",
     "build": "webpack --mode production"
   },
   "repository": {
@@ -33,10 +33,12 @@
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "^6.2.0",
+    "duplicate-package-checker-webpack-plugin": "^3.0.0",
     "file-loader": "^6.2.0",
     "html-loader": "^2.1.2",
     "html-webpack-plugin": "^5.3.2",
     "mini-css-extract-plugin": "^2.2.0",
+    "svgo-loader": "^3.0.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.48.0",
     "webpack-cli": "^4.7.2",
